ticket 8 refactored
